requirements.txt
flask==2.0.1
opencv-python==4.5.3.56
tensorflow==2.7.0
google-api-python-client==2.86.0
python-dotenv==0.19.0
numpy>=1.19.5